Refactor store & EventsView
