+++ packages/agents-ui/src/components/AgentDetails/EventsView/index.tsx
@@ -1,0 +1,1 @@
+export { EventsView } from "./EventsView";
